Create examples with env variables (#15)
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

### Issue Link, if available

### Description
We need to create our examples with environment variables set for
`DEX_ENDPOINT`. There was a KMS permission boundary needed where we need
to use our own KMS key to add environment variables to the function.
I've created the underlying infrastructure in the GitHub account for
this and verified it's working.

### Demo/Screenshots
See the GitHub run -> "Deploy Function". You can see the environment
variables are being added to the function properly now.

### Checklist

- [X] I have filled out every section of the PR template
- [X] I have thoroughly tested this change