cli: pull durable execution history (#17)
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

### Issue Link, if available

### Description
Updating the integration test workflow to pull the execution history
events for the example. We'll use this in a subsequent change to assert
the example has the expected history.

### Demo/Screenshots
See GitHub action.

### Checklist
- [X] I have filled out every section of the PR template
- [X] I have thoroughly tested this change