ci: add GitHub action to clean log groups created by integration tests (#55)
*Issue #, if available:*

*Description of changes:*
Adding a new GH action that will clean the log groups that get created
by running integration tests against each PR.

I'm not 100% sure this will work but I think so based on
https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#running-your-pull_request-workflow-when-a-pull-request-merges.
We'll know for sure when I merge it and it hits the webhook!


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.