refactor(testing-sdk): change operation storage to register operations on base class and use API interfaces (#89)
*Issue #, if available:*

DAR-SDK-305

*Description of changes:*

Refactoring operation storage to use API interfaces, and to register
operations on the base class.

This will allow for the cloud runner to call APIs such as
SendCallbackSuccess, and to register operations before the cloud runner
starts.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.