ci: increase integ test timeout, only cleanup on failure (#115)
*Issue #, if available:*

*Description of changes:*

Tests are timing out after 30 seconds. Increasing the timeout to 1
minute, and also changing the workflow to only cleanup when it's not
failed. This will allow retrying the integration tests to work better.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.