fix(testing-sdk): do not accept checkpoints or re-invocations after execution is completed (#269)
*Issue #, if available:*

*Description of changes:*

The language SDK seems to sometimes schedule checkpoints after the
execution is completed. This causes hanging tests in some cases. We
shouldn't be accepting these checkpoints in the testing library.

Sometimes, the checkpoints still come after the execution completes and
they're accepted by the testing library. Since the testing library is in
the same thread it doesn't get the execution completion before the
delayed checkpoints every time. Adding a 5xx on re-invocations in this
case to prevent re-invocations of the handler.

To properly fix this, it likely needs some deeper fixes with refactoring
of the testing library and moving things to different workers.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.