feat(testing-sdk): add support for invocation event (#268)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/189

*Description of changes:*

Adding support for InvocationCompleted event in the testing library.

Now:
- The cloud runner will return the correct result for getInvocations()
- The local runner will populate the InvocationCompleted event in the
history
- Invocations are no longer associated with operations (i.e., you can't
do getOperations on an invocation anymore) -> the history API does not
show this correspondence. It's possible that all operations between
InvocationCompleted events are from that invocation, but that needs to
be tested more thoroughly.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.