fix(sdk): remove singleton pattern from checkpoint system (#327)
**Changes**
- Remove Checkpoint Singleton Architecture
- ESLint plugin configuration from monorepo root to examples package
only
- ESLint plugin build issues (empty dist folder)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

**Important Note**
Checkpoint, Termination, DurableExecution, pendingOperation and many
other objects need access to oher one and although I address some of the
issues including circular dependencies and right order to create those,
there are still room to improve this.
Ideally all the data and operations that are for the entire execution
should be inside DurableExecution and operation and data related to
DurableContext and child context should be in DurableContext.
But this PR is getting too big and I'll address remaining issue in next
PRs