fix(sdk): fix waitForCallback determinism and serdes errors (#328)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/324

*Description of changes:*

Fixing waitForCallback serdes not working properly. The serdes was only
passed to the `createCallback` result, and `runInChildContext` had the
default serdes (JSON stringify/parse).

Before:
- Any deserialization that happened in createCallback would be lost on
replay, since runInChildContext had the default serdes
- If the serdes passed into createCallback created a string that could
not be serialized or deserialized into JSON, it would fail the execution
- For example, if you deserialized into a structure that had a circular
object, it would fail the execution (even though it's valid to do this)
- waitForCallback and createCallback both required the serdes to include
a `serialization` function, even though it was not used

After:
- Deserialization works both before and after replay
- Now, createCallback is not passed a serdes, and and runInChildContext
uses passThroughSerdes. Deserialization happens on the runInChildContext
result
- We can't pass serdes to runInChildContext since deserialization does
not happen on the initial execution, only on replay. We want this
deserialization to happen on the createCallback string result every time
- Deserialization into circular structures will not fail the execution
- waitForCallback and createCallback only use the `deserialization`
function of serdes.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.