fix(testing-sdk): reject waitForData promises when execution completes (#318)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/317

*Description of changes:*

Preventing hanging promises if execution completes while waitForData is
pending. Before, waitForData would never resolve/reject when the
execution completes and the operation was not found. Now, it will reject
all of the pending promises when it completes.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.