ci: remove version when syncing file (#363)
*Issue #, if available:*

*Description of changes:*

Currently syncing the SDK is not working since incrementing the package
version breaks the sync.

We can't sync with the version in the filename, since we would have to
parse the version from the filename to be able to untar it. We can just
sync without the version in the filename.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.